cleanup and fix validations
